refactor runtime into switchboard and focused modules
